\name{DelayDiffForR-package}
\alias{DelayDiffForR-package}
\alias{DelayDiffForR}
\docType{package}
\title{
\packageTitle{DelayDiffForR}
}
\description{
\packageDescription{DelayDiffForR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DelayDiffForR}
\packageIndices{DelayDiffForR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DelayDiffForR}

Maintainer: \packageMaintainer{DelayDiffForR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
